pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = ""
authors = ["nataliadudina"]
readme = "README.md"
packages = [{include = "coursework_5"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
